Sprites dos conexiones casi

+++ Practica2/Casilla.py
@@ -12,7 +12,7 @@ orientaciones = {
     (0, -1): 0,
     (-1, 0): 1,
     (0, 1): 2,
-    (1, 0): 3    
+    (1, 0): 3
 }
 
 
@@ -69,10 +69,10 @@ class Casilla(object):
     def orientacion(self):
         orientacion_media = 0
         for conexion in self.conexiones:
-            orientacion = orientaciones[conexion]
+            orientacion = orientaciones[conexion]            
             orientacion_media += orientacion
 
-        return int(orientacion_media / len(self.conexiones))
+        return orientacion_media / len(self.conexiones)
 
     def esta_vacia(self):
         return self.tipo == TiposCasilla.VACIA
